Atlassian only allows Jira backups every few days, fallback to existing one
Atlassian only allows the creation of a backup every couple of hours
(for me Jira only every 2 days). If someone triggers a backup (either
via UI or API), the script would raise an Exception due to a returned
code 412. However, it is possible to fetch the task ID of the last
backup and download that instead which is what this change implements.

Signed-off-by: Daniel Egger <daniel.egger@standards-consulting.de>